requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
dj-database-url==2.3.0
Django==5.1.6
django-coinpayments==0.1.4
django-cors-headers==4.4.0
django-model-utils==4.5.1
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.7
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
psycopg==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pyparsing==3.1.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.9.0
